[PT-SB-iOS] Add build variant for Tokoy/Multi region server
- Scheme file has added for different variant
-Config file added for different variant

- ref: https://sourcenext.backlog.com/view/PT_SK-8821

+++ PockeTalk/PockeTalk/PoceketalkSoftBank/HomePage/ViewController/SpeechProcessingViewController.swift
@@ -237,7 +237,7 @@ class SpeechProcessingViewController: BaseViewController{
             guard let `self` = self else { return }
             PrintUtility.printLog(tag: "Frequency : ", text: "\(avgVal)")
             let frequency = (avgVal >= 0.4) ? 0.9 : 0.5
-            self.speechProcessingVM.setFrequency(frequency)
+            self.speechProcessingVM.setFrequency(CGFloat(frequency))
         }
         
         service?.recordDidStop = { [weak self]  in
